refactor(PAT-64): Modularize. Create TicketMenu.
